backend/requirements.txt
Flask
flask-cors
pyjwt
bcrypt
python-dotenv
gunicorn